Move reviews under the site namespace
This seems more logical
